Moved theme code to a separate module named 'theme'.
